requirements.txt
fastapi==0.95.1
uvicorn==0.22.0  
python-dotenv==1.0.0
httpx==0.24.0
pydantic==1.10.7
numpy==1.24.3
opencv-python==4.7.0.72
python-multipart==0.0.6
numpy>=1.24.0
scipy>=1.10.0
pyodbc==4.0.39